remove unnecessary vm
