+++ lib/component/crop_tool_corner.dart
@@ -1,11 +1,26 @@
 
-import 'package:open_scanner/component/crop_tool_corner_vm.dart';
+import 'dart:math';
 import 'package:flutter/material.dart';
+import 'package:open_scanner/repo/crop_tool.dart';
 
 class CropToolCorner extends StatefulWidget {
-  final CropToolCornerVM vm;
+  final int index;
+  final int cornerIndex;
+  final CropToolRepo cropToolRepo;
 
-  const CropToolCorner({super.key, required this.vm});
+  Point<double> getCornerPosition() {
+    return cropToolRepo.getCropToolCorner(index, cornerIndex);
+  }
+
+  void moveCorner(Offset delta) {
+    return cropToolRepo.moveCorner(index, cornerIndex, delta);
+  }
+
+  listenCornerChanges(VoidCallback cb) {
+    return cropToolRepo.listenCropTool(index, cb);
+  }
+
+  const CropToolCorner(this.index, this.cornerIndex, this.cropToolRepo, {super.key});
 
   @override
   CropToolCornerState createState() => CropToolCornerState();
@@ -15,19 +30,19 @@ class CropToolCornerState extends State<CropToolCorner> {
   @override
   void initState() {
     super.initState();
-    widget.vm.listenCornerChanges(() => setState(() => {}));
+    widget.listenCornerChanges(() => setState(() => {}));
   }
 
   @override
   Widget build(BuildContext context) {
-    final cornerPosition = widget.vm.getCornerPosition();
+    final cornerPosition = widget.getCornerPosition();
     return Positioned(
       top: cornerPosition.y - 25,
       left: cornerPosition.x - 25,
       child: GestureDetector(
         onPanUpdate: (details) {
           setState(() {
-            widget.vm.moveCorner(details.delta);
+            widget.moveCorner(details.delta);
           });
         },
         child: Container(
@@ -37,7 +52,7 @@ class CropToolCornerState extends State<CropToolCorner> {
               shape: BoxShape.circle,
               color: const Color.fromRGBO(0, 255, 0, 0.5),
               border: Border.all(width: 1, color: Colors.red)),
-          child: Text(widget.vm.cornerIndex.toString()),
+          child: Text(widget.cornerIndex.toString()),
         ),
       ),
     );
