+++ lib/component/crop_tool_plane.dart
@@ -1,10 +1,25 @@
-import 'package:open_scanner/component/crop_tool_plane_vm.dart';
+import 'dart:math';
 import 'package:flutter/material.dart';
+import 'package:open_scanner/component/crop_tool_plane_painter.dart';
+import 'package:open_scanner/repo/crop_tool.dart';
 
 class CropToolPlane extends StatefulWidget {
-  final CropToolPlaneVM vm;
+  final int index;
+  final CropToolRepo cropToolRepo;
 
-  const CropToolPlane({super.key, required this.vm});
+  listenCornerChanges(VoidCallback cb) {
+    return cropToolRepo.listenCropTool(index, cb);
+  }
+
+  void moveCropTool(Offset delta) {
+    return cropToolRepo.moveCropTool(index, delta);
+  }
+
+  List<Point<double>> getCropToolCorners() {
+    return cropToolRepo.getCropToolCorners(index);
+  }
+
+  const CropToolPlane(this.index, this.cropToolRepo, {super.key});
 
   @override
   CropToolPlaneState createState() => CropToolPlaneState();
@@ -14,7 +29,7 @@ class CropToolPlaneState extends State<CropToolPlane> {
   @override
   void initState() {
     super.initState();
-    widget.vm.listenCornerChanges(() => setState(() => {}));
+    widget.listenCornerChanges(() => setState(() => {}));
   }
 
   @override
@@ -22,45 +37,13 @@ class CropToolPlaneState extends State<CropToolPlane> {
     return GestureDetector(
       onPanUpdate: (details) {
         setState(() {
-          widget.vm.moveCropTool(details.delta);
+          widget.moveCropTool(details.delta);
         });
       },
       child: CustomPaint(
-        painter: CropToolPlanePainter(widget.vm),
+        painter: CropToolPlanePainter(widget),
         child: Container(),
       ),
     );
-  }
-}
-
-class CropToolPlanePainter extends CustomPainter {
-  late Path path;
-
-  CropToolPlanePainter(CropToolPlaneVM vm) {
-    final corners = vm.getCropToolCorners();
-    path = Path()
-      ..moveTo(corners[0].x, corners[0].y)
-      ..lineTo(corners[1].x, corners[1].y)
-      ..lineTo(corners[3].x, corners[3].y)
-      ..lineTo(corners[2].x, corners[2].y);
-  }
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    var paint = Paint()
-      ..color = const Color.fromRGBO(255, 0, 0, 0.3)
-      ..style = PaintingStyle.fill;
-
-    canvas.drawPath(path, paint);
-  }
-
-  @override
-  bool shouldRepaint(covariant CustomPainter oldDelegate) {
-    return true;
-  }
-
-  @override
-  bool hitTest(Offset position) {
-    return path.contains(position);
   }
 }
