--- lib/component/crop_tool_plane_vm.dart
@@ -1,24 +1,0 @@
-
-import 'dart:math';
-
-import 'package:open_scanner/repo/crop_tool.dart';
-import 'package:flutter/material.dart';
-
-class CropToolPlaneVM {
-  late int index;
-  late CropToolRepo cropToolRepo;
-
-  CropToolPlaneVM(this.index, this.cropToolRepo);
-
-  listenCornerChanges(VoidCallback cb) {
-    return cropToolRepo.listenCropTool(index, cb);
-  }
-
-  void moveCropTool(Offset delta) {
-    return cropToolRepo.moveCropTool(index, delta);
-  }
-
-  List<Point<double>> getCropToolCorners() {
-    return cropToolRepo.getCropToolCorners(index);
-  }
-}
